Refactoring `ExtSimAdapterData`
