+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased/Snapshot]
 
+### Changed
+- Updated the maven central publishing scripts [#339](https://github.com/ie3-institute/simonaAPI/issues/339)
+- Refactoring `ExtSimAdapterData` [#347](https://github.com/ie3-institute/simonaAPI/issues/347)
+
+## [0.10.0] - 2025-09-10
+
 ### Added
 - Added support for external em communication [#304](https://github.com/ie3-institute/simonaAPI/issues/304)
