+++ src/main/java/edu/ie3/simona/api/ExtLinkInterface.java
@@ -6,8 +6,8 @@
 
 package edu.ie3.simona.api;
 
+import edu.ie3.simona.api.data.ExtSimAdapterData;
 import edu.ie3.simona.api.exceptions.NoExtSimulationException;
-import edu.ie3.simona.api.simulation.ExtSimAdapterData;
 import edu.ie3.simona.api.simulation.ExtSimulation;
 
 /**
