+++ src/main/java/edu/ie3/simona/api/simulation/ExtSimulation.java
@@ -6,6 +6,7 @@
 
 package edu.ie3.simona.api.simulation;
 
+import edu.ie3.simona.api.data.ExtSimAdapterData;
 import edu.ie3.simona.api.data.connection.ExtDataConnection;
 import edu.ie3.simona.api.ontology.simulation.*;
 import java.util.Optional;
