+++ src/test/groovy/edu/ie3/simona/api/simulation/ExtSimulationSpec.groovy
@@ -1,5 +1,6 @@
 package edu.ie3.simona.api.simulation
 
+import edu.ie3.simona.api.data.ExtSimAdapterData
 import edu.ie3.simona.api.data.connection.ExtDataConnection
 import edu.ie3.simona.api.ontology.simulation.ActivationMessage
 import edu.ie3.simona.api.ontology.simulation.CompletionMessage
@@ -70,7 +71,7 @@ class ExtSimulationSpec extends Specification {
         def tick = -1L
         def newTick = 0L
         def extSimAdapter = testKit.createTestProbe(ControlResponseMessageFromExt)
-        def extSimData = new ExtSimAdapterData(extSimAdapter.ref(), new String[0])
+        def extSimData = new ExtSimAdapterData(extSimAdapter.ref(), new String[0], null, null)
         def extSim = new TestSimulation(newTick, Optional.of(-2L))
         extSim.setAdapterData(extSimData)
 
@@ -86,7 +87,7 @@ class ExtSimulationSpec extends Specification {
     def "An ExtSimulation should handle activation and return given new triggers"() {
         given:
         def extSimAdapter = testKit.createTestProbe(ControlResponseMessageFromExt)
-        def extSimData = new ExtSimAdapterData(extSimAdapter.ref(), new String[0])
+        def extSimData = new ExtSimAdapterData(extSimAdapter.ref(), new String[0], null, null)
         def newTickOpt = newTick.isEmpty() ?
                     Optional.<Long>empty() : Optional.of(newTick.first())
         def extSim = new TestSimulation(-2L, newTickOpt)
@@ -111,7 +112,7 @@ class ExtSimulationSpec extends Specification {
     def "An ExtSimulation should handle termination properly"() {
         given:
         def extSimAdapter = testKit.createTestProbe(ControlResponseMessageFromExt)
-        def extSimData = new ExtSimAdapterData(extSimAdapter.ref(), new String[0])
+        def extSimData = new ExtSimAdapterData(extSimAdapter.ref(), new String[0], null, null)
         def extSim = new TestSimulation(-1L, Optional.empty())
         extSim.setAdapterData(extSimData)
 
@@ -134,7 +135,7 @@ class ExtSimulationSpec extends Specification {
     def "An ExtSimulation should handle unknown messages by throwing an exception"() {
         given:
         def extSimAdapter = testKit.createTestProbe(ControlResponseMessageFromExt)
-        def extSimData = new ExtSimAdapterData(extSimAdapter.ref(), new String[0])
+        def extSimData = new ExtSimAdapterData(extSimAdapter.ref(), new String[0], null, null)
         def extSim = new TestSimulation(-1L, Optional.empty())
         extSim.setAdapterData(extSimData)
